Pull out some slow deferred tests to quality
