+++ hypothesis-python/tests/cover/test_deferred_strategies.py
@@ -33,13 +33,6 @@ def test_binary_tree():
     assert minimal(tree, lambda x: isinstance(x, tuple)) == (0, 0)
 
 
-def test_large_branching_tree():
-    tree = st.deferred(
-        lambda: st.integers() | st.tuples(tree, tree, tree, tree, tree))
-    assert minimal(tree) == 0
-    assert minimal(tree, lambda x: isinstance(x, tuple)) == (0,) * 5
-
-
 def test_mutual_recursion():
     t = st.deferred(lambda: a | b)
     a = st.deferred(lambda: st.none() | st.tuples(st.just('a'), b))
@@ -48,25 +41,6 @@ def test_mutual_recursion():
     for c in ('a', 'b'):
         assert minimal(
             t, lambda x: x is not None and x[0] == c) == (c, None)
-
-
-def test_non_trivial_json():
-    json = st.deferred(
-        lambda: st.none() | st.floats() | st.text() | lists | objects
-    )
-
-    lists = st.lists(json)
-    objects = st.dictionaries(st.text(), json)
-
-    assert minimal(json) is None
-
-    small_list = minimal(json, lambda x: isinstance(x, list) and x)
-    assert small_list == [None]
-
-    x = minimal(
-        json, lambda x: isinstance(x, dict) and isinstance(x.get(''), list))
-
-    assert x == {'': []}
 
 
 def test_errors_on_non_function_define():
@@ -125,13 +99,6 @@ def test_mutually_recursive_tuples_draw_nothing():
 
     assert_no_examples(x)
     assert_no_examples(y)
-
-
-def test_self_recursive_lists():
-    x = st.deferred(lambda: st.lists(x))
-    assert minimal(x) == []
-    assert minimal(x, bool) == [[]]
-    assert minimal(x, lambda x: len(x) > 1) == [[], []]
 
 
 def test_literals_strategy_is_valid():
